requirements.txt
pytest
requests
inquirer
Pillow
mock
tinydb
